+++ tests/components/bluetooth/test_models.py
@@ -1,63 +1,28 @@
 """Tests for the Bluetooth integration models."""
 from __future__ import annotations
 
-from datetime import timedelta
-import time
 from unittest.mock import patch
 
 import bleak
-from bleak import BleakClient, BleakError
+from bleak import BleakError
 from bleak.backends.device import BLEDevice
 from bleak.backends.scanner import AdvertisementData
 import pytest
 
 from homeassistant.components.bluetooth import (
-    BaseHaRemoteScanner,
     BaseHaScanner,
     HaBleakScannerWrapper,
     HaBluetoothConnector,
 )
-from homeassistant.components.bluetooth.const import (
-    CONNECTABLE_FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS,
-    FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS,
-)
 from homeassistant.components.bluetooth.wrappers import HaBleakClientWrapper
-import homeassistant.util.dt as dt_util
 
 from . import (
+    MockBleakClient,
     _get_manager,
     generate_advertisement_data,
     inject_advertisement,
     inject_advertisement_with_source,
 )
-
-from tests.common import async_fire_time_changed
-
-
-class MockBleakClient(BleakClient):
-    """Mock bleak client."""
-
-    def __init__(self, *args, **kwargs):
-        """Mock init."""
-        super().__init__(*args, **kwargs)
-        self._device_path = "/dev/test"
-
-    @property
-    def is_connected(self) -> bool:
-        """Mock connected."""
-        return True
-
-    async def connect(self, *args, **kwargs):
-        """Mock connect."""
-        return True
-
-    async def disconnect(self, *args, **kwargs):
-        """Mock disconnect."""
-        pass
-
-    async def get_services(self, *args, **kwargs):
-        """Mock get_services."""
-        return []
 
 
 async def test_wrapped_bleak_scanner(hass, enable_bluetooth):
@@ -354,251 +319,4 @@ async def test_ble_device_with_proxy_client_out_of_connections_uses_best_availab
     assert client.is_connected is True
     client.set_disconnected_callback(lambda client: None)
     await client.disconnect()
-    cancel()
-
-
-async def test_remote_scanner(hass):
-    """Test the remote scanner base class merges advertisement_data."""
-    manager = _get_manager()
-
-    switchbot_device = BLEDevice(
-        "44:44:33:11:23:45",
-        "wohand",
-        {},
-        rssi=-100,
-    )
-    switchbot_device_adv = generate_advertisement_data(
-        local_name="wohand",
-        service_uuids=["050a021a-0000-1000-8000-00805f9b34fb"],
-        service_data={"050a021a-0000-1000-8000-00805f9b34fb": b"\n\xff"},
-        manufacturer_data={1: b"\x01"},
-        rssi=-100,
-    )
-    switchbot_device_2 = BLEDevice(
-        "44:44:33:11:23:45",
-        "w",
-        {},
-        rssi=-100,
-    )
-    switchbot_device_adv_2 = generate_advertisement_data(
-        local_name="wohand",
-        service_uuids=["00000001-0000-1000-8000-00805f9b34fb"],
-        service_data={"00000001-0000-1000-8000-00805f9b34fb": b"\n\xff"},
-        manufacturer_data={1: b"\x01", 2: b"\x02"},
-        rssi=-100,
-    )
-
-    class FakeScanner(BaseHaRemoteScanner):
-        def inject_advertisement(
-            self, device: BLEDevice, advertisement_data: AdvertisementData
-        ) -> None:
-            """Inject an advertisement."""
-            self._async_on_advertisement(
-                device.address,
-                advertisement_data.rssi,
-                device.name,
-                advertisement_data.service_uuids,
-                advertisement_data.service_data,
-                advertisement_data.manufacturer_data,
-                advertisement_data.tx_power,
-            )
-
-    new_info_callback = manager.scanner_adv_received
-    connector = (
-        HaBluetoothConnector(MockBleakClient, "mock_bleak_client", lambda: False),
-    )
-    scanner = FakeScanner(hass, "esp32", new_info_callback, connector, True)
-    scanner.async_setup()
-    cancel = manager.async_register_scanner(scanner, True)
-
-    scanner.inject_advertisement(switchbot_device, switchbot_device_adv)
-
-    data = scanner.discovered_devices_and_advertisement_data
-    discovered_device, discovered_adv_data = data[switchbot_device.address]
-    assert discovered_device.address == switchbot_device.address
-    assert discovered_device.name == switchbot_device.name
-    assert (
-        discovered_adv_data.manufacturer_data == switchbot_device_adv.manufacturer_data
-    )
-    assert discovered_adv_data.service_data == switchbot_device_adv.service_data
-    assert discovered_adv_data.service_uuids == switchbot_device_adv.service_uuids
-    scanner.inject_advertisement(switchbot_device_2, switchbot_device_adv_2)
-
-    data = scanner.discovered_devices_and_advertisement_data
-    discovered_device, discovered_adv_data = data[switchbot_device.address]
-    assert discovered_device.address == switchbot_device.address
-    assert discovered_device.name == switchbot_device.name
-    assert discovered_adv_data.manufacturer_data == {1: b"\x01", 2: b"\x02"}
-    assert discovered_adv_data.service_data == {
-        "050a021a-0000-1000-8000-00805f9b34fb": b"\n\xff",
-        "00000001-0000-1000-8000-00805f9b34fb": b"\n\xff",
-    }
-    assert set(discovered_adv_data.service_uuids) == {
-        "050a021a-0000-1000-8000-00805f9b34fb",
-        "00000001-0000-1000-8000-00805f9b34fb",
-    }
-
-    cancel()
-
-
-async def test_remote_scanner_expires_connectable(hass):
-    """Test the remote scanner expires stale connectable data."""
-    manager = _get_manager()
-
-    switchbot_device = BLEDevice(
-        "44:44:33:11:23:45",
-        "wohand",
-        {},
-        rssi=-100,
-    )
-    switchbot_device_adv = generate_advertisement_data(
-        local_name="wohand",
-        service_uuids=[],
-        manufacturer_data={1: b"\x01"},
-        rssi=-100,
-    )
-
-    class FakeScanner(BaseHaRemoteScanner):
-        def inject_advertisement(
-            self, device: BLEDevice, advertisement_data: AdvertisementData
-        ) -> None:
-            """Inject an advertisement."""
-            self._async_on_advertisement(
-                device.address,
-                advertisement_data.rssi,
-                device.name,
-                advertisement_data.service_uuids,
-                advertisement_data.service_data,
-                advertisement_data.manufacturer_data,
-                advertisement_data.tx_power,
-            )
-
-    new_info_callback = manager.scanner_adv_received
-    connector = (
-        HaBluetoothConnector(MockBleakClient, "mock_bleak_client", lambda: False),
-    )
-    scanner = FakeScanner(hass, "esp32", new_info_callback, connector, True)
-    scanner.async_setup()
-    cancel = manager.async_register_scanner(scanner, True)
-
-    start_time_monotonic = time.monotonic()
-    scanner.inject_advertisement(switchbot_device, switchbot_device_adv)
-
-    devices = scanner.discovered_devices
-    assert len(scanner.discovered_devices) == 1
-    assert len(scanner.discovered_devices_and_advertisement_data) == 1
-    assert devices[0].name == "wohand"
-
-    expire_monotonic = (
-        start_time_monotonic
-        + CONNECTABLE_FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS
-        + 1
-    )
-    expire_utc = dt_util.utcnow() + timedelta(
-        seconds=CONNECTABLE_FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS + 1
-    )
-    with patch(
-        "homeassistant.components.bluetooth.models.MONOTONIC_TIME",
-        return_value=expire_monotonic,
-    ):
-        async_fire_time_changed(hass, expire_utc)
-        await hass.async_block_till_done()
-
-    devices = scanner.discovered_devices
-    assert len(scanner.discovered_devices) == 0
-    assert len(scanner.discovered_devices_and_advertisement_data) == 0
-
-    cancel()
-
-
-async def test_remote_scanner_expires_non_connectable(hass):
-    """Test the remote scanner expires stale non connectable data."""
-    manager = _get_manager()
-
-    switchbot_device = BLEDevice(
-        "44:44:33:11:23:45",
-        "wohand",
-        {},
-        rssi=-100,
-    )
-    switchbot_device_adv = generate_advertisement_data(
-        local_name="wohand",
-        service_uuids=[],
-        manufacturer_data={1: b"\x01"},
-        rssi=-100,
-    )
-
-    class FakeScanner(BaseHaRemoteScanner):
-        def inject_advertisement(
-            self, device: BLEDevice, advertisement_data: AdvertisementData
-        ) -> None:
-            """Inject an advertisement."""
-            self._async_on_advertisement(
-                device.address,
-                advertisement_data.rssi,
-                device.name,
-                advertisement_data.service_uuids,
-                advertisement_data.service_data,
-                advertisement_data.manufacturer_data,
-                advertisement_data.tx_power,
-            )
-
-    new_info_callback = manager.scanner_adv_received
-    connector = (
-        HaBluetoothConnector(MockBleakClient, "mock_bleak_client", lambda: False),
-    )
-    scanner = FakeScanner(hass, "esp32", new_info_callback, connector, False)
-    scanner.async_setup()
-    cancel = manager.async_register_scanner(scanner, True)
-
-    start_time_monotonic = time.monotonic()
-    scanner.inject_advertisement(switchbot_device, switchbot_device_adv)
-
-    devices = scanner.discovered_devices
-    assert len(scanner.discovered_devices) == 1
-    assert len(scanner.discovered_devices_and_advertisement_data) == 1
-    assert devices[0].name == "wohand"
-
-    assert (
-        FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS
-        > CONNECTABLE_FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS
-    )
-
-    # The connectable timeout is not used for non connectable devices
-    expire_monotonic = (
-        start_time_monotonic
-        + CONNECTABLE_FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS
-        + 1
-    )
-    expire_utc = dt_util.utcnow() + timedelta(
-        seconds=CONNECTABLE_FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS + 1
-    )
-    with patch(
-        "homeassistant.components.bluetooth.models.MONOTONIC_TIME",
-        return_value=expire_monotonic,
-    ):
-        async_fire_time_changed(hass, expire_utc)
-        await hass.async_block_till_done()
-
-    assert len(scanner.discovered_devices) == 1
-    assert len(scanner.discovered_devices_and_advertisement_data) == 1
-
-    # The non connectable timeout is used for non connectable devices
-    # which is always longer than the connectable timeout
-    expire_monotonic = (
-        start_time_monotonic + FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS + 1
-    )
-    expire_utc = dt_util.utcnow() + timedelta(
-        seconds=FALLBACK_MAXIMUM_STALE_ADVERTISEMENT_SECONDS + 1
-    )
-    with patch(
-        "homeassistant.components.bluetooth.models.MONOTONIC_TIME",
-        return_value=expire_monotonic,
-    ):
-        async_fire_time_changed(hass, expire_utc)
-        await hass.async_block_till_done()
-
-    assert len(scanner.discovered_devices) == 0
-    assert len(scanner.discovered_devices_and_advertisement_data) == 0
-
     cancel()
